refactor CreateSignature, add issuer tests, refactor signature files
